task(GUIDEFRAME-36): Rename files to include GuideFrame in name
