+++ guideframe_assembly.py
