+++ guideframe_utils.py
@@ -1,5 +1,5 @@
-from audio import generate_voicover
-from video import start_ffmpeg_recording, stop_ffmpeg_recording
+from guideframe_audio import generate_voicover
+from guideframe_video import start_ffmpeg_recording, stop_ffmpeg_recording
 import time
 import sys
