+++ selenium_automation_tests.py
@@ -1,4 +1,4 @@
-from selenium_functions import *
+from guideframe_selenium import *
 import time
 from guideframe_utils import get_env_settings  # Importing the guide_step and get_env_settings functions from guideframe_utils.py
