+++ tutors_test.py
@@ -1,5 +1,5 @@
-from selenium_functions import * # Moved all selenium functions to external file
-from assembly import assemble  # Importing the assemble_clips function from assembly.py
+from guideframe_selenium import * # Moved all selenium functions to external file
+from guideframe_assembly import assemble  # Importing the assemble_clips function from assembly.py
 from guideframe_utils import guide_step, get_env_settings  # Importing the guide_step and get_env_settings functions from guideframe_utils.py
